requirements.txt
argcomplete==1.10.0
beautifulsoup4==4.8.0
bokeh==2.0.2
chardet==3.0.4
DAWG-Python==0.7.2
docopt==0.6.2
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
IMAPClient==2.1.0
Jinja2==2.11.2
joblib==0.14.1
llvmlite==0.32.0
lxml==4.5.0
MarkupSafe==1.1.1
numba==0.49.0
numpy==1.18.3
olefile==0.46
packaging==20.3
pdfminer.six==20181108
Pillow==7.1.2
pycryptodome==3.9.7
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pyparsing==2.4.7
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2019.3
PyYAML==5.3.1
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.12.0
sklearn==0.0
sortedcontainers==2.1.0
soupsieve==2.0
SpeechRecognition==3.8.1
tbb==2019.0
textract==1.6.3
tornado==6.0.4
typing-extensions==3.7.4.2
tzlocal==1.5.1
umap-learn==0.4.1
xlrd==1.2.0
XlsxWriter==1.2.8
